requirements/requirements.txt
asgiref==3.6.0
attrs==22.2.0
autopep8==2.0.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
dj-rest-auth==3.0.0
Django==4.2
django-debug-toolbar==4.0.0
django-js-asset==2.0.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.1
idna==3.4
inflection==0.5.1
jsonschema==4.17.3
paypalrestsdk==1.13.1
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.1.1
pyrsistent==0.19.3
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
requests==2.30.0
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.2
